Numerous updates
I cleaned up the .gitignore and __INIT__.py files.

I added functionality for creating the backup-YYYYMMDD-hm folders which
all the backed up files shall be put within.
The for-loop on line 91 now funcitons properly. I have this whole
function commented out however since it will be superflous if I won't
make use of the regex I previously wrote.
The regex on line 98 is now removed.
I also removed the uneccesary for-loop on line 65.