Numerous Changes
I have cleaned up backupy 1.0 and 1.1 directories. I also added the 2.0
directory and work on backupy 2.0 has now begun.
I eliminated a glitch in 1.0 and 1.1 that resulted in the user having to
run the script twice to make a backup.

Extensive work has been made on the linux installation part of the
SETUP.py